Moving the community issues workflow out of the .github directory
Follow up on https://github.com/tenstorrent/tt-github-actions/issues/52
specifically moving this up and out, and starting the move around on
where these live.  This will move it up and out, next commit here will
be to copy it back so I'm not breaking anything while I adjust all the
repos that are calling into this.  Once all are updated the copy in
.github/workflows/ can go away

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>