Community issue labeling: adding some debugging and clarify the 'on'
Mostly adding some minor debugging on status if we do not find 204 for
the return status, as well as turning off what triggers this, as we
should be relying on the workflows in caller, not here, to get things
labeled / working.  Even if we go to use it here we should be using an
extra caller, not modifying this.

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>